@@ -14,12 +14,14 @@ declare(strict_types=1);
 
 namespace FOD\DBALClickHouse;
 
+use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\TrimMode;
 use Doctrine\DBAL\Schema\ForeignKeyConstraint;
 use Doctrine\DBAL\Schema\Index;
 use Doctrine\DBAL\Schema\TableDiff;
+use Doctrine\DBAL\Schema\UniqueConstraint;
 use Doctrine\DBAL\Types\BigIntType;
 use Doctrine\DBAL\Types\BlobType;
 use Doctrine\DBAL\Types\DateTimeType;
@@ -34,8 +36,9 @@ use Doctrine\DBAL\Types\Type;
 use FOD\DBALClickHouse\Types\BitNumericalClickHouseType;
 use FOD\DBALClickHouse\Types\DatableClickHouseType;
 use FOD\DBALClickHouse\Types\NumericalClickHouseType;
-use FOD\DBALClickHouse\Types\StringClickHouseType;
+use FOD\DBALClickHouse\Types\StringableClickHouseType;
 use FOD\DBALClickHouse\Types\UnsignedNumericalClickHouseType;
+
 use function addslashes;
 use function array_filter;
 use function array_key_exists;
@@ -48,13 +51,10 @@ use function func_get_args;
 use function get_class;
 use function implode;
 use function in_array;
+use function mb_stripos;
 use function sprintf;
-use function stripos;
 use function trim;
 
-/**
- * Provides the behavior, features and SQL dialect of the ClickHouse database platform.
- */
 class ClickHousePlatform extends AbstractPlatform
 {
     protected const TIME_MINUTE = 60;
@@ -65,128 +65,128 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getBooleanTypeDeclarationSQL(array $columnDef) : string
+    public function getBooleanTypeDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
             UnsignedNumericalClickHouseType::UNSIGNED_CHAR .
             NumericalClickHouseType::TYPE_INT . BitNumericalClickHouseType::EIGHT_BIT,
-            $columnDef
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getIntegerTypeDeclarationSQL(array $columnDef) : string
+    public function getIntegerTypeDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
-            $this->_getCommonIntegerTypeDeclarationSQL($columnDef) .
+            $this->_getCommonIntegerTypeDeclarationSQL($column) .
             NumericalClickHouseType::TYPE_INT . BitNumericalClickHouseType::THIRTY_TWO_BIT,
-            $columnDef
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getBigIntTypeDeclarationSQL(array $columnDef) : string
+    public function getBigIntTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $columnDef);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getSmallIntTypeDeclarationSQL(array $columnDef) : string
+    public function getSmallIntTypeDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
-            $this->_getCommonIntegerTypeDeclarationSQL($columnDef) .
+            $this->_getCommonIntegerTypeDeclarationSQL($column) .
             NumericalClickHouseType::TYPE_INT . BitNumericalClickHouseType::SIXTEEN_BIT,
-            $columnDef
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function _getCommonIntegerTypeDeclarationSQL(array $columnDef) : string
+    protected function _getCommonIntegerTypeDeclarationSQL(array $column): string
     {
-        if (! empty($columnDef['autoincrement'])) {
-            throw ClickHouseException::notSupported('Clickhouse do not support AUTO_INCREMENT fields');
+        if (!empty($column['autoincrement'])) {
+            throw Exception::notSupported('Clickhouse does not support AUTO_INCREMENT fields');
         }
 
-        return empty($columnDef['unsigned']) ? '' : UnsignedNumericalClickHouseType::UNSIGNED_CHAR;
+        return empty($column['unsigned']) ? '' : UnsignedNumericalClickHouseType::UNSIGNED_CHAR;
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function initializeDoctrineTypeMappings() : void
+    protected function initializeDoctrineTypeMappings(): void
     {
         $this->doctrineTypeMapping = [
-            'int8' => 'smallint',
-            'int16' => 'integer',
-            'int32' => 'integer',
-            'int64' => 'bigint',
-            'uint8' => 'smallint',
-            'uint16' => 'integer',
-            'uint32' => 'integer',
-            'uint64' => 'bigint',
+            'int8'    => 'smallint',
+            'int16'   => 'integer',
+            'int32'   => 'integer',
+            'int64'   => 'bigint',
+            'uint8'   => 'smallint',
+            'uint16'  => 'integer',
+            'uint32'  => 'integer',
+            'uint64'  => 'bigint',
             'float32' => 'decimal',
             'float64' => 'float',
 
-            'string' => 'string',
+            'string'      => 'string',
             'fixedstring' => 'string',
-            'date' => 'date',
-            'datetime' => 'datetime',
+            'date'        => 'date',
+            'datetime'    => 'datetime',
 
-            'array(int8)' => 'array',
-            'array(int16)' => 'array',
-            'array(int32)' => 'array',
-            'array(int64)' => 'array',
-            'array(uint8)' => 'array',
-            'array(uint16)' => 'array',
-            'array(uint32)' => 'array',
-            'array(uint64)' => 'array',
+            'array(int8)'    => 'array',
+            'array(int16)'   => 'array',
+            'array(int32)'   => 'array',
+            'array(int64)'   => 'array',
+            'array(uint8)'   => 'array',
+            'array(uint16)'  => 'array',
+            'array(uint32)'  => 'array',
+            'array(uint64)'  => 'array',
             'array(float32)' => 'array',
             'array(float64)' => 'array',
 
-            'array(string)' => 'array',
-            'array(date)' => 'array',
+            'array(string)'   => 'array',
+            'array(date)'     => 'array',
             'array(datetime)' => 'array',
 
-            'enum8' => 'string',
+            'enum8'  => 'string',
             'enum16' => 'string',
 
-            'nullable(int8)' => 'smallint',
-            'nullable(int16)' => 'integer',
-            'nullable(int32)' => 'integer',
-            'nullable(int64)' => 'bigint',
-            'nullable(uint8)' => 'smallint',
-            'nullable(uint16)' => 'integer',
-            'nullable(uint32)' => 'integer',
-            'nullable(uint64)' => 'bigint',
+            'nullable(int8)'    => 'smallint',
+            'nullable(int16)'   => 'integer',
+            'nullable(int32)'   => 'integer',
+            'nullable(int64)'   => 'bigint',
+            'nullable(uint8)'   => 'smallint',
+            'nullable(uint16)'  => 'integer',
+            'nullable(uint32)'  => 'integer',
+            'nullable(uint64)'  => 'bigint',
             'nullable(float32)' => 'decimal',
             'nullable(float64)' => 'float',
 
-            'nullable(string)' => 'string',
+            'nullable(string)'      => 'string',
             'nullable(fixedstring)' => 'string',
-            'nullable(date)' => 'date',
-            'nullable(datetime)' => 'datetime',
+            'nullable(date)'        => 'date',
+            'nullable(datetime)'    => 'datetime',
 
-            'array(nullable(int8))' => 'array',
-            'array(nullable(int16))' => 'array',
-            'array(nullable(int32))' => 'array',
-            'array(nullable(int64))' => 'array',
-            'array(nullable(uint8))' => 'array',
-            'array(nullable(uint16))' => 'array',
-            'array(nullable(uint32))' => 'array',
-            'array(nullable(uint64))' => 'array',
+            'array(nullable(int8))'    => 'array',
+            'array(nullable(int16))'   => 'array',
+            'array(nullable(int32))'   => 'array',
+            'array(nullable(int64))'   => 'array',
+            'array(nullable(uint8))'   => 'array',
+            'array(nullable(uint16))'  => 'array',
+            'array(nullable(uint32))'  => 'array',
+            'array(nullable(uint64))'  => 'array',
             'array(nullable(float32))' => 'array',
             'array(nullable(float64))' => 'array',
 
-            'array(nullable(string))' => 'array',
-            'array(nullable(date))' => 'array',
+            'array(nullable(string))'   => 'array',
+            'array(nullable(date))'     => 'array',
             'array(nullable(datetime))' => 'array',
         ];
     }
@@ -194,74 +194,74 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    protected function getVarcharTypeDeclarationSQLSnippet($length, $fixed) : string
+    protected function getVarcharTypeDeclarationSQLSnippet($length, $fixed): string
     {
         return $fixed
-            ? (StringClickHouseType::TYPE_FIXED_STRING . '(' . $length . ')')
-            : StringClickHouseType::TYPE_STRING;
+            ? (StringableClickHouseType::TYPE_FIXED_STRING . '(' . $length . ')')
+            : StringableClickHouseType::TYPE_STRING;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getVarcharTypeDeclarationSQL(array $field) : string
+    public function getVarcharTypeDeclarationSQL(array $column): string
     {
-        if (! isset($field['length'])) {
-            $field['length'] = $this->getVarcharDefaultLength();
+        if (!isset($column['length'])) {
+            $column['length'] = $this->getVarcharDefaultLength();
         }
 
-        $fixed = $field['fixed'] ?? false;
+        $fixed = $column['fixed'] ?? false;
 
         $maxLength = $fixed
             ? $this->getCharMaxLength()
             : $this->getVarcharMaxLength();
 
-        if ($field['length'] > $maxLength) {
-            return $this->getClobTypeDeclarationSQL($field);
+        if ($column['length'] > $maxLength) {
+            return $this->getClobTypeDeclarationSQL($column);
         }
 
         return $this->prepareDeclarationSQL(
-            $this->getVarcharTypeDeclarationSQLSnippet($field['length'], $fixed),
-            $field
+            $this->getVarcharTypeDeclarationSQLSnippet($column['length'], $fixed),
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getBinaryTypeDeclarationSQLSnippet($length, $fixed) : string
+    protected function getBinaryTypeDeclarationSQLSnippet($length, $fixed): string
     {
-        return StringClickHouseType::TYPE_STRING;
+        return StringableClickHouseType::TYPE_STRING;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getClobTypeDeclarationSQL(array $field) : string
+    public function getClobTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $field);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getBlobTypeDeclarationSQL(array $field) : string
+    public function getBlobTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $field);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getName() : string
+    public function getName(): string
     {
-        return 'clickhouse';
+        return 'ClickHouse';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getIdentifierQuoteCharacter() : string
+    public function getIdentifierQuoteCharacter(): string
     {
         return '`';
     }
@@ -269,7 +269,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getVarcharDefaultLength() : int
+    public function getVarcharDefaultLength(): int
     {
         return 512;
     }
@@ -277,7 +277,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCountExpression($column) : string
+    public function getCountExpression($column): string
     {
         return 'COUNT()';
     }
@@ -287,7 +287,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getMd5Expression($column) : string
+    public function getMd5Expression($column): string
     {
         return 'MD5(CAST(' . $column . ' AS String))';
     }
@@ -295,7 +295,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getLengthExpression($column) : string
+    public function getLengthExpression($column): string
     {
         return 'lengthUTF8(CAST(' . $column . ' AS String))';
     }
@@ -303,7 +303,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getSqrtExpression($column) : string
+    public function getSqrtExpression($column): string
     {
         return 'sqrt(' . $column . ')';
     }
@@ -311,7 +311,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getRoundExpression($column, $decimals = 0) : string
+    public function getRoundExpression($column, $decimals = 0): string
     {
         return 'round(' . $column . ', ' . $decimals . ')';
     }
@@ -319,7 +319,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getModExpression($expression1, $expression2) : string
+    public function getModExpression($expression1, $expression2): string
     {
         return 'modulo(' . $expression1 . ', ' . $expression2 . ')';
     }
@@ -327,17 +327,14 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getTrimExpression($str, $pos = TrimMode::UNSPECIFIED, $char = false) : string
+    public function getTrimExpression($str, $mode = TrimMode::UNSPECIFIED, $char = false): string
     {
-        if (! $char) {
-            switch ($pos) {
-                case TrimMode::LEADING:
-                    return $this->getLtrimExpression($str);
-                case TrimMode::TRAILING:
-                    return $this->getRtrimExpression($str);
-                default:
-                    return sprintf("replaceRegexpAll(%s, '(^\\\s+|\\\s+$)', '')", $str);
-            }
+        if (!$char) {
+            return match ($mode) {
+                TrimMode::LEADING => $this->getLtrimExpression($str),
+                TrimMode::TRAILING => $this->getRtrimExpression($str),
+                default => sprintf("replaceRegexpAll(%s, '(^\\\s+|\\\s+$)', '')", $str),
+            };
         }
 
         return sprintf("replaceRegexpAll(%s, '(^%s+|%s+$)', '')", $str, $char, $char);
@@ -346,7 +343,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getRtrimExpression($str) : string
+    public function getRtrimExpression($str): string
     {
         return sprintf("replaceRegexpAll(%s, '(\\\s+$)', '')", $str);
     }
@@ -354,7 +351,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getLtrimExpression($str) : string
+    public function getLtrimExpression($str): string
     {
         return sprintf("replaceRegexpAll(%s, '(^\\\s+)', '')", $str);
     }
@@ -362,7 +359,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getUpperExpression($str) : string
+    public function getUpperExpression($str): string
     {
         return 'upperUTF8(' . $str . ')';
     }
@@ -370,7 +367,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getLowerExpression($str) : string
+    public function getLowerExpression($str): string
     {
         return 'lowerUTF8(' . $str . ')';
     }
@@ -378,7 +375,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getLocateExpression($str, $substr, $startPos = false) : string
+    public function getLocateExpression($str, $substr, $startPos = false): string
     {
         return 'positionUTF8(' . $str . ', ' . $substr . ')';
     }
@@ -386,7 +383,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getNowExpression() : string
+    public function getNowExpression(): string
     {
         return 'now()';
     }
@@ -394,19 +391,19 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getSubstringExpression($value, $from, $length = null) : string
+    public function getSubstringExpression($string, $start, $length = null): string
     {
         if ($length === null) {
             throw new \InvalidArgumentException("'length' argument must be a constant");
         }
 
-        return 'substringUTF8(' . $value . ', ' . $from . ', ' . $length . ')';
+        return 'substringUTF8(' . $string . ', ' . $start . ', ' . $length . ')';
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getConcatExpression() : string
+    public function getConcatExpression(): string
     {
         return 'concat(' . implode(', ', func_get_args()) . ')';
     }
@@ -414,7 +411,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getIsNullExpression($expression) : string
+    public function getIsNullExpression($expression): string
     {
         return 'isNull(' . $expression . ')';
     }
@@ -422,7 +419,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getIsNotNullExpression($expression) : string
+    public function getIsNotNullExpression($expression): string
     {
         return 'isNotNull(' . $expression . ')';
     }
@@ -430,7 +427,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getAcosExpression($value) : string
+    public function getAcosExpression($value): string
     {
         return 'acos(' . $value . ')';
     }
@@ -438,7 +435,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getSinExpression($value) : string
+    public function getSinExpression($value): string
     {
         return 'sin(' . $value . ')';
     }
@@ -446,7 +443,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getPiExpression() : string
+    public function getPiExpression(): string
     {
         return 'pi()';
     }
@@ -454,7 +451,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCosExpression($value) : string
+    public function getCosExpression($value): string
     {
         return 'cos(' . $value . ')';
     }
@@ -462,7 +459,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateDiffExpression($date1, $date2) : string
+    public function getDateDiffExpression($date1, $date2): string
     {
         return 'CAST(' . $date1 . ' AS Date) - CAST(' . $date2 . ' AS Date)';
     }
@@ -470,7 +467,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateAddSecondsExpression($date, $seconds) : string
+    public function getDateAddSecondsExpression($date, $seconds): string
     {
         return $date . ' + ' . $seconds;
     }
@@ -478,7 +475,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateSubSecondsExpression($date, $seconds) : string
+    public function getDateSubSecondsExpression($date, $seconds): string
     {
         return $date . ' - ' . $seconds;
     }
@@ -486,7 +483,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateAddMinutesExpression($date, $minutes) : string
+    public function getDateAddMinutesExpression($date, $minutes): string
     {
         return $date . ' + ' . $minutes * self::TIME_MINUTE;
     }
@@ -494,7 +491,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateSubMinutesExpression($date, $minutes) : string
+    public function getDateSubMinutesExpression($date, $minutes): string
     {
         return $date . ' - ' . $minutes * self::TIME_MINUTE;
     }
@@ -502,7 +499,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateAddHourExpression($date, $hours) : string
+    public function getDateAddHourExpression($date, $hours): string
     {
         return $date . ' + ' . $hours * self::TIME_HOUR;
     }
@@ -510,7 +507,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateSubHourExpression($date, $hours) : string
+    public function getDateSubHourExpression($date, $hours): string
     {
         return $date . ' - ' . $hours * self::TIME_HOUR;
     }
@@ -518,7 +515,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateAddDaysExpression($date, $days) : string
+    public function getDateAddDaysExpression($date, $days): string
     {
         return $date . ' + ' . $days * self::TIME_DAY;
     }
@@ -526,7 +523,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateSubDaysExpression($date, $days) : string
+    public function getDateSubDaysExpression($date, $days): string
     {
         return $date . ' - ' . $days * self::TIME_DAY;
     }
@@ -534,7 +531,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateAddWeeksExpression($date, $weeks) : string
+    public function getDateAddWeeksExpression($date, $weeks): string
     {
         return $date . ' + ' . $weeks * self::TIME_WEEK;
     }
@@ -542,7 +539,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDateSubWeeksExpression($date, $weeks) : string
+    public function getDateSubWeeksExpression($date, $weeks): string
     {
         return $date . ' - ' . $weeks * self::TIME_WEEK;
     }
@@ -550,7 +547,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getBitAndComparisonExpression($value1, $value2) : string
+    public function getBitAndComparisonExpression($value1, $value2): string
     {
         return 'bitAnd(' . $value1 . ', ' . $value2 . ')';
     }
@@ -558,7 +555,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getBitOrComparisonExpression($value1, $value2) : string
+    public function getBitOrComparisonExpression($value1, $value2): string
     {
         return 'bitOr(' . $value1 . ', ' . $value2 . ')';
     }
@@ -566,81 +563,81 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getForUpdateSQL() : string
+    public function getForUpdateSQL(): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function appendLockHint($fromClause, $lockMode) : string
+    public function appendLockHint($fromClause, $lockMode): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getReadLockSQL() : string
+    public function getReadLockSQL(): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getWriteLockSQL() : string
+    public function getWriteLockSQL(): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDropIndexSQL($index, $table = null) : string
+    public function getDropIndexSQL($index, $table = null): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDropConstraintSQL($constraint, $table) : string
+    public function getDropConstraintSQL($constraint, $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDropForeignKeySQL($foreignKey, $table) : string
+    public function getDropForeignKeySQL($foreignKey, $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getCommentOnColumnSQL($tableName, $columnName, $comment) : string
+    public function getCommentOnColumnSQL($tableName, $columnName, $comment): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function _getCreateTableSQL($tableName, array $columns, array $options = []) : array
+    protected function _getCreateTableSQL($name, array $columns, array $options = []): array
     {
-        $engine        = ! empty($options['engine']) ? $options['engine'] : 'ReplacingMergeTree';
+        $engine        = !empty($options['engine']) ? $options['engine'] : 'ReplacingMergeTree';
         $engineOptions = '';
 
-        if (isset($options['uniqueConstraints']) && ! empty($options['uniqueConstraints'])) {
-            throw ClickHouseException::notSupported('uniqueConstraints');
+        if (isset($options['uniqueConstraints']) && !empty($options['uniqueConstraints'])) {
+            throw Exception::notSupported('uniqueConstraints');
         }
 
-        if (isset($options['indexes']) && ! empty($options['indexes'])) {
-            throw ClickHouseException::notSupported('uniqueConstraints');
+        if (isset($options['indexes']) && !empty($options['indexes'])) {
+            throw Exception::notSupported('uniqueConstraints');
         }
 
         /**
@@ -658,34 +655,36 @@ class ClickHousePlatform extends AbstractPlatform
             ],
             true
         )) {
-            $indexGranularity   = ! empty($options['indexGranularity']) ? $options['indexGranularity'] : 8192;
+            $indexGranularity   = !empty($options['indexGranularity']) ? $options['indexGranularity'] : 8192;
             $samplingExpression = '';
 
             /**
              * eventDateColumn section
              */
             $dateColumnParams = [
-                'type' => Type::getType('date'),
+                'type'    => Type::getType('date'),
                 'default' => 'today()',
             ];
-            if (! empty($options['eventDateProviderColumn'])) {
+
+            if (!empty($options['eventDateProviderColumn'])) {
                 $options['eventDateProviderColumn'] = trim($options['eventDateProviderColumn']);
-                if (! isset($columns[$options['eventDateProviderColumn']])) {
+
+                if (!isset($columns[$options['eventDateProviderColumn']])) {
                     throw new \Exception(
-                        'Table `' . $tableName . '` has not column with name: `' . $options['eventDateProviderColumn']
+                        'Table `' . $name . '` has not column with name: `' . $options['eventDateProviderColumn']
                     );
                 }
 
-                if (! ($columns[$options['eventDateProviderColumn']]['type'] instanceof DateType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof DateTimeType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof TextType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof IntegerType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof SmallIntType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof BigIntType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof FloatType) &&
-                    ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof DecimalType) &&
+                if (!($columns[$options['eventDateProviderColumn']]['type'] instanceof DateType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof DateTimeType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof TextType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof IntegerType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof SmallIntType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof BigIntType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof FloatType) &&
+                    !($columns[$options['eventDateProviderColumn']]['type'] instanceof DecimalType) &&
                     (
-                        ! ($columns[$options['eventDateProviderColumn']]['type'] instanceof StringType) ||
+                        !($columns[$options['eventDateProviderColumn']]['type'] instanceof StringType) ||
                         $columns[$options['eventDateProviderColumn']]['fixed']
                     )
                 ) {
@@ -703,14 +702,13 @@ class ClickHousePlatform extends AbstractPlatform
                         ('toDate(toDateTime(' . $options['eventDateProviderColumn'] . '))') :
                         ('toDate(' . $options['eventDateProviderColumn'] . ')');
             }
+
             if (empty($options['eventDateColumn'])) {
-                $dateColumns = array_filter($columns, function ($column) {
-                    return $column['type'] instanceof DateType;
-                });
+                $dateColumns = array_filter($columns, fn (array $column): bool => $column['type'] instanceof DateType);
 
                 if ($dateColumns) {
                     throw new \Exception(
-                        'Table `' . $tableName . '` has DateType columns: `' . implode(
+                        'Table `' . $name . '` has DateType columns: `' . implode(
                             '`, `',
                             array_keys($dateColumns)
                         ) .
@@ -721,9 +719,9 @@ class ClickHousePlatform extends AbstractPlatform
 
                 $eventDateColumnName = 'EventDate';
             } elseif (isset($columns[$options['eventDateColumn']])) {
-                if (! ($columns[$options['eventDateColumn']]['type'] instanceof DateType)) {
+                if (!($columns[$options['eventDateColumn']]['type'] instanceof DateType)) {
                     throw new \Exception(
-                        'In table `' . $tableName . '` you have set field `' .
+                        'In table `' . $name . '` you have set field `' .
                         $options['eventDateColumn'] .
                         '` (' . get_class($columns[$options['eventDateColumn']]['type']) . ')
                          as `eventDateColumn`, but it is not instance of DateType'
@@ -731,10 +729,12 @@ class ClickHousePlatform extends AbstractPlatform
                 }
 
                 $eventDateColumnName = $options['eventDateColumn'];
+
                 unset($columns[$options['eventDateColumn']]);
             } else {
                 $eventDateColumnName = $options['eventDateColumn'];
             }
+
             $dateColumnParams['name'] = $eventDateColumnName;
             // insert into very beginning
             $columns = [$eventDateColumnName => $dateColumnParams] + $columns;
@@ -747,7 +747,8 @@ class ClickHousePlatform extends AbstractPlatform
             }
 
             $primaryIndex = array_values($options['primary']);
-            if (! empty($options['samplingExpression'])) {
+
+            if (!empty($options['samplingExpression'])) {
                 $samplingExpression = ', ' . $options['samplingExpression'];
                 $primaryIndex[]     = $options['samplingExpression'];
             }
@@ -766,19 +767,19 @@ class ClickHousePlatform extends AbstractPlatform
             /**
              * any specific MergeTree* table parameters
              */
-            if ($engine === 'ReplacingMergeTree' && ! empty($options['versionColumn'])) {
-                if (! isset($columns[$options['versionColumn']])) {
+            if ($engine === 'ReplacingMergeTree' && !empty($options['versionColumn'])) {
+                if (!isset($columns[$options['versionColumn']])) {
                     throw new \Exception(
                         'If you specify `versionColumn` for ReplacingMergeTree table -- 
                         you must add this column manually (any of UInt*, Date or DateTime types)'
                     );
                 }
 
-                if (! $columns[$options['versionColumn']]['type'] instanceof IntegerType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof BigIntType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof SmallIntType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof DateType &&
-                    ! $columns[$options['versionColumn']]['type'] instanceof DateTimeType
+                if (!$columns[$options['versionColumn']]['type'] instanceof IntegerType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof BigIntType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof SmallIntType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof DateType &&
+                    !$columns[$options['versionColumn']]['type'] instanceof DateTimeType
                 ) {
                     throw new \Exception(
                         'For ReplacingMergeTree tables `versionColumn` must be any of UInt* family, 
@@ -795,7 +796,7 @@ class ClickHousePlatform extends AbstractPlatform
 
         $sql[] = sprintf(
             'CREATE TABLE %s (%s) ENGINE = %s%s',
-            $tableName,
+            $name,
             $this->getColumnDeclarationListSQL($columns),
             $engine,
             $engineOptions
@@ -807,20 +808,21 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCreateForeignKeySQL(ForeignKeyConstraint $foreignKey, $table) : string
+    public function getCreateForeignKeySQL(ForeignKeyConstraint $foreignKey, $table): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getAlterTableSQL(TableDiff $diff) : array
+    public function getAlterTableSQL(TableDiff $diff): array
     {
         $columnSql  = [];
         $queryParts = [];
-        if ($diff->newName !== false || ! empty($diff->renamedColumns)) {
-            throw ClickHouseException::notSupported('RENAME COLUMN');
+
+        if ($diff->newName !== false || !empty($diff->renamedColumns)) {
+            throw Exception::notSupported('RENAME COLUMN');
         }
 
         foreach ($diff->addedColumns as $column) {
@@ -857,19 +859,16 @@ class ClickHousePlatform extends AbstractPlatform
             }
 
             $queryParts[] = 'MODIFY COLUMN ' . $this->getColumnDeclarationSQL(
-                $column->getQuotedName($this),
-                $columnArray
-            );
+                    $column->getQuotedName($this),
+                    $columnArray
+                );
         }
 
         $sql      = [];
         $tableSql = [];
 
-        if (! $this->onSchemaAlterTable($diff, $tableSql) && (count($queryParts) > 0)) {
-            $sql[] = 'ALTER TABLE ' . $diff->getName($this)->getQuotedName($this) . ' ' . implode(
-                ', ',
-                $queryParts
-            );
+        if (!$this->onSchemaAlterTable($diff, $tableSql) && (count($queryParts) > 0)) {
+            $sql[] = 'ALTER TABLE ' . $diff->getName($this)->getQuotedName($this) . ' ' . implode(', ', $queryParts);
         }
 
         return array_merge($sql, $tableSql, $columnSql);
@@ -878,36 +877,28 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    protected function getPreAlterTableIndexForeignKeySQL(TableDiff $diff) : array
+    protected function getPreAlterTableIndexForeignKeySQL(TableDiff $diff): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getPostAlterTableIndexForeignKeySQL(TableDiff $diff) : array
+    protected function getPostAlterTableIndexForeignKeySQL(TableDiff $diff): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getRenameIndexSQL($oldIndexName, Index $index, $tableName) : array
+    protected function getRenameIndexSQL($oldIndexName, Index $index, $tableName): array
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    protected function _getAlterTableIndexForeignKeySQL(TableDiff $diff) : array
-    {
-        throw ClickHouseException::notSupported(__METHOD__);
-    }
-
-    protected function prepareDeclarationSQL(string $declarationSQL, array $columnDef) : string
+    protected function prepareDeclarationSQL(string $declarationSQL, array $columnDef): string
     {
         if (array_key_exists('notnull', $columnDef) && $columnDef['notnull'] === false) {
             return 'Nullable(' . $declarationSQL . ')';
@@ -919,14 +910,14 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getColumnDeclarationSQL($name, array $field) : string
+    public function getColumnDeclarationSQL($name, array $column): string
     {
-        if (isset($field['columnDefinition'])) {
-            $columnDef = $this->getCustomTypeDeclarationSQL($field);
+        if (isset($column['columnDefinition'])) {
+            $columnDef = $this->getCustomTypeDeclarationSQL($column);
         } else {
-            $default = $this->getDefaultValueDeclarationSQL($field);
+            $default = $this->getDefaultValueDeclarationSQL($column);
 
-            $columnDef = $field['type']->getSqlDeclaration($field, $this) . $default;
+            $columnDef = $column['type']->getSqlDeclaration($column, $this) . $default;
         }
 
         return $name . ' ' . $columnDef;
@@ -935,80 +926,79 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDecimalTypeDeclarationSQL(array $columnDef) : string
+    public function getDecimalTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $columnDef);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getCheckDeclarationSQL(array $definition) : string
+    public function getCheckDeclarationSQL(array $definition): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getUniqueConstraintDeclarationSQL($name, Index $index) : string
+    public function getUniqueConstraintDeclarationSQL($name, UniqueConstraint $constraint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getIndexDeclarationSQL($name, Index $index) : string
+    public function getIndexDeclarationSQL($name, Index $index): string
     {
-        // Index declaration in statements like CREATE TABLE is not supported.
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getForeignKeyDeclarationSQL(ForeignKeyConstraint $foreignKey) : string
+    public function getForeignKeyDeclarationSQL(ForeignKeyConstraint $foreignKey): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getAdvancedForeignKeyOptionsSQL(ForeignKeyConstraint $foreignKey) : string
+    public function getAdvancedForeignKeyOptionsSQL(ForeignKeyConstraint $foreignKey): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getForeignKeyReferentialActionSQL($action) : string
+    public function getForeignKeyReferentialActionSQL($action): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getForeignKeyBaseDeclarationSQL(ForeignKeyConstraint $foreignKey) : string
+    public function getForeignKeyBaseDeclarationSQL(ForeignKeyConstraint $foreignKey): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getUniqueFieldDeclarationSQL() : string
+    public function getUniqueFieldDeclarationSQL(): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getCurrentDateSQL() : string
+    public function getCurrentDateSQL(): string
     {
         return 'today()';
     }
@@ -1016,7 +1006,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCurrentTimeSQL() : string
+    public function getCurrentTimeSQL(): string
     {
         return 'now()';
     }
@@ -1024,7 +1014,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCurrentTimestampSQL() : string
+    public function getCurrentTimestampSQL(): string
     {
         return 'toUnixTimestamp(now())';
     }
@@ -1032,7 +1022,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getListDatabasesSQL() : string
+    public function getListDatabasesSQL(): string
     {
         return 'SHOW DATABASES';
     }
@@ -1040,7 +1030,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getListTableColumnsSQL($table, $database = null) : string
+    public function getListTableColumnsSQL($table, $database = null): string
     {
         return sprintf(
             'DESCRIBE TABLE %s',
@@ -1051,7 +1041,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getListTablesSQL() : string
+    public function getListTablesSQL(): string
     {
         return "SELECT database, name FROM system.tables WHERE database != 'system' AND engine != 'View'";
     }
@@ -1059,7 +1049,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getListViewsSQL($database) : string
+    public function getListViewsSQL($database): string
     {
         return "SELECT name FROM system.tables WHERE database != 'system' AND engine = 'View'";
     }
@@ -1067,7 +1057,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCreateViewSQL($name, $sql) : string
+    public function getCreateViewSQL($name, $sql): string
     {
         return 'CREATE VIEW ' . $this->quoteIdentifier($name) . ' AS ' . $sql;
     }
@@ -1075,7 +1065,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDropViewSQL($name) : string
+    public function getDropViewSQL($name): string
     {
         return 'DROP TABLE ' . $this->quoteIdentifier($name);
     }
@@ -1083,60 +1073,60 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getCreateDatabaseSQL($database) : string
+    public function getCreateDatabaseSQL($name): string
     {
-        return 'CREATE DATABASE ' . $database;
+        return 'CREATE DATABASE ' . $name;
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateTimeTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getDateTimeTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateTimeTzTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getDateTimeTzTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE_TIME, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getTimeTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getTimeTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(StringClickHouseType::TYPE_STRING, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(StringableClickHouseType::TYPE_STRING, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDateTypeDeclarationSQL(array $fieldDeclaration) : string
+    public function getDateTypeDeclarationSQL(array $column): string
     {
-        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE, $fieldDeclaration);
+        return $this->prepareDeclarationSQL(DatableClickHouseType::TYPE_DATE, $column);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getFloatDeclarationSQL(array $fieldDeclaration) : string
+    public function getFloatDeclarationSQL(array $column): string
     {
         return $this->prepareDeclarationSQL(
             NumericalClickHouseType::TYPE_FLOAT . BitNumericalClickHouseType::SIXTY_FOUR_BIT,
-            $fieldDeclaration
+            $column
         );
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getDefaultTransactionIsolationLevel() : int
+    public function getDefaultTransactionIsolationLevel(): int
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /* supports*() methods */
@@ -1144,7 +1134,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function supportsTransactions() : bool
+    public function supportsTransactions(): bool
     {
         return false;
     }
@@ -1152,7 +1142,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function supportsSavepoints() : bool
+    public function supportsSavepoints(): bool
     {
         return false;
     }
@@ -1160,7 +1150,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function supportsPrimaryConstraints() : bool
+    public function supportsPrimaryConstraints(): bool
     {
         return false;
     }
@@ -1168,7 +1158,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function supportsForeignKeyConstraints() : bool
+    public function supportsForeignKeyConstraints(): bool
     {
         return false;
     }
@@ -1176,7 +1166,7 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function supportsGettingAffectedRows() : bool
+    public function supportsGettingAffectedRows(): bool
     {
         return false;
     }
@@ -1184,13 +1174,14 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    protected function doModifyLimitQuery($query, $limit, $offset) : string
+    protected function doModifyLimitQuery($query, $limit, $offset): string
     {
         if ($limit === null) {
             return $query;
         }
 
         $query .= ' LIMIT ';
+
         if ($offset !== null) {
             $query .= $offset . ', ';
         }
@@ -1203,15 +1194,15 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getEmptyIdentityInsertSQL($tableName, $identifierColumnName) : string
+    public function getEmptyIdentityInsertSQL($quotedTableName, $quotedIdentifierColumnName): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function getTruncateTableSQL($tableName, $cascade = false) : string
+    public function getTruncateTableSQL($tableName, $cascade = false): string
     {
         /**
          * For MergeTree* engines may be done with next workaround:
@@ -1219,37 +1210,37 @@ class ClickHousePlatform extends AbstractPlatform
          * SELECT partition FROM system.parts WHERE table= '$tableName';
          * ALTER TABLE $tableName DROP PARTITION $partitionName
          */
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function createSavePoint($savepoint) : string
+    public function createSavePoint($savepoint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function releaseSavePoint($savepoint) : string
+    public function releaseSavePoint($savepoint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function rollbackSavePoint($savepoint) : string
+    public function rollbackSavePoint($savepoint): string
     {
-        throw ClickHouseException::notSupported(__METHOD__);
+        throw Exception::notSupported(__METHOD__);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getReservedKeywordsClass() : string
+    protected function getReservedKeywordsClass(): string
     {
         return ClickHouseKeywords::class;
     }
@@ -1257,13 +1248,15 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDefaultValueDeclarationSQL($field) : string
+    public function getDefaultValueDeclarationSQL($column): string
     {
-        if (! isset($field['default'])) {
+        if (!isset($column['default'])) {
             return '';
         }
-        $defaultValue = $this->quoteStringLiteral($field['default']);
-        $fieldType    = $field['type'] ?: null;
+
+        $defaultValue = $this->quoteStringLiteral($column['default']);
+        $fieldType    = $column['type'] ?: null;
+
         if ($fieldType !== null) {
             if ($fieldType === DatableClickHouseType::TYPE_DATE ||
                 $fieldType instanceof DateType ||
@@ -1277,9 +1270,9 @@ class ClickHousePlatform extends AbstractPlatform
                     && Type::getType('BigInt')->getBindingType() === ParameterType::INTEGER
                 )
             ) {
-                $defaultValue = $field['default'];
+                $defaultValue = $column['default'];
             } elseif ($fieldType === DatableClickHouseType::TYPE_DATE_TIME &&
-                $field['default'] === $this->getCurrentTimestampSQL()
+                $column['default'] === $this->getCurrentTimestampSQL()
             ) {
                 $defaultValue = $this->getCurrentTimestampSQL();
             }
@@ -1291,29 +1284,27 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function getDoctrineTypeMapping($dbType) : string
+    public function getDoctrineTypeMapping($dbType): string
     {
-        // FixedString
-        if (stripos($dbType, 'fixedstring') === 0) {
+        if (mb_stripos($dbType, 'fixedstring') === 0) {
             $dbType = 'fixedstring';
         }
 
-        //Enum8
-        if (stripos($dbType, 'enum8') === 0) {
+        if (mb_stripos($dbType, 'enum8') === 0) {
             $dbType = 'enum8';
         }
 
-        //Enum16
-        if (stripos($dbType, 'enum16') === 0) {
+        if (mb_stripos($dbType, 'enum16') === 0) {
             $dbType = 'enum16';
         }
+
         return parent::getDoctrineTypeMapping($dbType);
     }
 
     /**
      * {@inheritDoc}
      */
-    public function quoteStringLiteral($str) : string
+    public function quoteStringLiteral($str): string
     {
         $c = $this->getStringLiteralQuoteCharacter();
 
@@ -1323,10 +1314,18 @@ class ClickHousePlatform extends AbstractPlatform
     /**
      * {@inheritDoc}
      */
-    public function quoteSingleIdentifier($str) : string
+    public function quoteSingleIdentifier($str): string
     {
         $c = $this->getIdentifierQuoteCharacter();
 
         return $c . addslashes($str) . $c;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getCurrentDatabaseExpression(): string
+    {
+        return 'DATABASE()';
     }
 }
