@@ -19,7 +19,7 @@ use Doctrine\DBAL\Schema\Column;
 use Doctrine\DBAL\Schema\Index;
 use Doctrine\DBAL\Schema\View;
 use Doctrine\DBAL\Types\Type;
-use const CASE_LOWER;
+
 use function array_change_key_case;
 use function array_filter;
 use function array_key_exists;
@@ -28,17 +28,15 @@ use function array_reverse;
 use function current;
 use function explode;
 use function is_array;
+use function mb_stripos;
+use function mb_strtolower;
 use function preg_match;
 use function preg_replace;
 use function str_replace;
-use function stripos;
-use function strpos;
-use function strtolower;
 use function trim;
 
-/**
- * Schema manager for the ClickHouse DBMS.
- */
+use const CASE_LOWER;
+
 class ClickHouseSchemaManager extends AbstractSchemaManager
 {
     /**
@@ -52,7 +50,7 @@ class ClickHouseSchemaManager extends AbstractSchemaManager
     /**
      * {@inheritdoc}
      */
-    protected function _getPortableViewDefinition($view) : View
+    protected function _getPortableViewDefinition($view): View
     {
         $statement = $this->_conn->fetchOne('SHOW CREATE TABLE ' . $view['name']);
 
@@ -62,7 +60,7 @@ class ClickHouseSchemaManager extends AbstractSchemaManager
     /**
      * {@inheritdoc}
      */
-    public function listTableIndexes($table) : array
+    public function listTableIndexes($table): array
     {
         $tableView = $this->_getPortableViewDefinition(['name' => $table]);
 
@@ -75,9 +73,7 @@ class ClickHouseSchemaManager extends AbstractSchemaManager
         if (is_array($matches) && array_key_exists(2, $matches)) {
             $indexColumns = array_filter(
                 array_map('trim', explode(',', $matches[2])),
-                function (string $column) {
-                    return strpos($column, '(') === false;
-                }
+                fn (string $column): bool => !str_contains($column, '(')
             );
 
             return [
@@ -96,7 +92,7 @@ class ClickHouseSchemaManager extends AbstractSchemaManager
     /**
      * {@inheritdoc}
      */
-    protected function _getPortableTableColumnDefinition($tableColumn) : Column
+    protected function _getPortableTableColumnDefinition($tableColumn): Column
     {
         $tableColumn = array_change_key_case($tableColumn, CASE_LOWER);
 
@@ -110,7 +106,7 @@ class ClickHouseSchemaManager extends AbstractSchemaManager
             $notnull    = false;
         }
 
-        if (stripos($columnType, 'fixedstring') === 0) {
+        if (mb_stripos($columnType, 'fixedstring') === 0) {
             // get length from FixedString definition
             $length = preg_replace('~.*\(([0-9]*)\).*~', '$1', $columnType);
             $dbType = 'fixedstring';
@@ -118,29 +114,30 @@ class ClickHouseSchemaManager extends AbstractSchemaManager
         }
 
         $unsigned = false;
-        if (stripos($columnType, 'uint') === 0) {
+
+        if (mb_stripos($columnType, 'uint') === 0) {
             $unsigned = true;
         }
 
-        if (! isset($tableColumn['name'])) {
+        if (!isset($tableColumn['name'])) {
             $tableColumn['name'] = '';
         }
 
         $default = null;
-        //TODO process not only DEFAULT type, but ALIAS and MATERIALIZED too
-        if ($tableColumn['default_expression'] && strtolower($tableColumn['default_type']) === 'default') {
+
+        // @todo process not only DEFAULT type, but ALIAS and MATERIALIZED too
+        if ($tableColumn['default_expression'] && mb_strtolower($tableColumn['default_type']) === 'default') {
             $default = $tableColumn['default_expression'];
         }
 
         $options = [
-            'length' => $length,
-            'notnull' => $notnull,
-            'default' => $default,
-            'primary' => false,
-            'fixed' => $fixed,
-            'unsigned' => $unsigned,
+            'length'        => $length,
+            'unsigned'      => $unsigned,
+            'fixed'         => $fixed,
+            'default'       => $default,
+            'notnull'       => $notnull,
             'autoincrement' => false,
-            'comment' => null,
+            'comment'       => null,
         ];
 
         return new Column(
