@@ -14,17 +14,14 @@ declare(strict_types=1);
 
 namespace FOD\DBALClickHouse\Types;
 
-/**
- * Array(Int16) Type
- */
 class ArrayInt16Type extends ArrayType implements BitNumericalClickHouseType
 {
-    public function getBits() : int
+    public function getBits(): int
     {
         return BitNumericalClickHouseType::SIXTEEN_BIT;
     }
 
-    public function getBaseClickHouseType() : string
+    public function getBaseClickHouseType(): string
     {
         return NumericalClickHouseType::TYPE_INT;
     }
