@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /*
  * This file is part of the FODDBALClickHouse package -- Doctrine DBAL library
  * for ClickHouse (a column-oriented DBMS for OLAP <https://clickhouse.yandex/>)
@@ -11,7 +14,9 @@
 
 namespace FOD\DBALClickHouse\Tests;
 
+use Doctrine\DBAL\Configuration;
 use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\DriverManager;
 use Doctrine\DBAL\Exception;
 use PHPUnit\Framework\TestCase;
 
@@ -22,41 +27,38 @@ use PHPUnit\Framework\TestCase;
  */
 class CreateConnectionTest extends TestCase
 {
-    public function testCreateConnectionWithRightParams()
+    public function testCreateConnectionWithRightParams(): void
     {
         $this->assertInstanceOf(Connection::class, self::createConnection());
     }
 
-    public function testCreateConnectionWithBadParams()
+    public function testCreateConnectionWithBadParams(): void
     {
         $this->expectException(Exception::class);
         $this->assertInstanceOf(Connection::class, self::createConnection([]));
     }
 
-    /**
-     * @param null|array $params
-     * @return Connection
-     */
-    public static function createConnection($params = null)
+    public static function createConnection(array $params = null): Connection
     {
-        if (null === $params) {
+        if ($params === null) {
             $params = [
-                'host' => phpunit_ch_host,
-                'port' => phpunit_ch_port,
-                'user' => phpunit_ch_user,
-                'password' => phpunit_ch_password,
-                'dbname' => phpunit_ch_dbname,
-                'driverClass' => phpunit_ch_driver_class,
-                'wrapperClass' => phpunit_ch_wrapper_class,
+                'host'          => phpunit_ch_host,
+                'port'          => phpunit_ch_port,
+                'user'          => phpunit_ch_user,
+                'password'      => phpunit_ch_password,
+                'dbname'        => phpunit_ch_dbname,
+                'driverClass'   => phpunit_ch_driver_class,
+                'wrapperClass'  => phpunit_ch_wrapper_class,
                 'driverOptions' => [
                     'extremes'                => false,
                     'readonly'                => true,
                     'max_execution_time'      => 30,
                     'enable_http_compression' => 0,
-                    'https'                   => false
+                    'https'                   => false,
                 ],
             ];
         }
-        return \Doctrine\DBAL\DriverManager::getConnection($params, new \Doctrine\DBAL\Configuration());
+
+        return DriverManager::getConnection($params, new Configuration());
     }
 }
